requirements.txt
anyio==3.6.1
autoflake==1.4
attrs==21.4.0
bandit==1.7.4
black==22.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
isort==5.10.1
iniconfig==1.1.1
joblib==1.1.0
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
sniffio==1.2.0
starlette==0.19.1
safety==2.0.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.11.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
boto3==1.24.30
